requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2
django-filter==23.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.7
iniconfig==2.0.0
install==1.3.5
packaging==24.1
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2024.1
requests==2.26.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.0
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
